ch04/myproject_virtualenv/src/django-myproject/requirements/_base.txt
coverage==4.5.3
Django==2.2.4
django-appconf==1.0.3
django-classy-tags==0.9.0
django-crispy-forms==1.7.2
django-imagekit==4.0.2
django-sekizai==1.0.0
pilkit==2.0
Pillow==6.1.0
psycopg2-binary==2.8.3
pytz==2019.2
six==1.12.0
sqlparse==0.3.0
